refactor remarks

+++ src/test/java/ru/practicum/yandex/user/CreateUserTest.java
@@ -1,6 +1,7 @@
 package ru.practicum.yandex.user;
 
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import ru.practicum.yandex.model.user.User;
@@ -15,37 +16,47 @@ public class CreateUserTest {
     User user;
     UserClient userClient;
 
-   @Before
+    boolean isDelete;
+    String accessToken;
+    boolean isCreate;
+
+    @Before
     public void start() {
-        user = User.getRegister();
+        user = User.getRandomDataRegister();
         userClient = new UserClient();
     }
+
+    @After
+    public void finish() {
+        if (accessToken != null) {
+            isDelete = userClient.delete(accessToken)
+                    .statusCode(SC_ACCEPTED).extract().path("success");
+            assertTrue(isDelete);
+        }
+    }
+
 
     @Test
     @DisplayName("Registration of a unique user ")
     public void createUniqueUserTest() {
-       boolean isCreate = userClient.create(user)
-               .statusCode(SC_OK).extract().path("success");
-       assertTrue(isCreate);
-       String accessToken = userClient.login(UserLogin.getLogin(user))
-               .statusCode(SC_OK).extract().path("accessToken");
-       boolean isDelete = userClient.delete(accessToken)
-               .statusCode(SC_ACCEPTED).extract().path("success");
-       assertTrue(isDelete);
+        isCreate = userClient.create(user)
+                .statusCode(SC_OK).extract().path("success");
+        assertTrue(isCreate);
+        accessToken = userClient.login(UserLogin.getLogin(user))
+                .statusCode(SC_OK).extract().path("accessToken");
+
     }
 
     @Test
     @DisplayName("Registration of a registered user")
     public void createSecondRegisteredUserTest() {
-       userClient.create(user)
-               .statusCode(SC_OK).extract().path("success");
-       String accessToken = userClient.login(UserLogin.getLogin(user))
-               .statusCode(SC_OK).extract().path("accessToken");
-       String error = userClient.create(user)
-               .statusCode(SC_FORBIDDEN).extract().path("message");
-       userClient.delete(accessToken)
-               .statusCode(SC_ACCEPTED).extract().path("success");
-       assertEquals("User already exists", error);
+        userClient.create(user)
+                .statusCode(SC_OK).extract().path("success");
+        String error = userClient.create(user)
+                .statusCode(SC_FORBIDDEN).extract().path("message");
+        assertEquals("User already exists", error);
+        accessToken = userClient.login(UserLogin.getLogin(user))
+                .statusCode(SC_OK).extract().path("accessToken");
     }
 
     @Test
@@ -55,6 +66,10 @@ public class CreateUserTest {
         String error = userClient.create(user)
                 .statusCode(SC_FORBIDDEN).extract().path("message");
         assertEquals("Email, password and name are required fields", error);
+        accessToken = userClient.login(UserLogin.getLogin(user))
+                .extract().path("accessToken");
 
-   }
+    }
+
+
 }
